#신입-베이스캠프-9기-루키토끼사랑/49 수정-float 변환

+++ main.c
@@ -55,6 +55,6 @@ int multi(int number1, int number2)
 int div(int number1, int number2)
 {
 	//TODO implementation
-	float result = number1 / number2;
+	float result = (float)number1 / (float)number2;
 	return result;
 }
